split apollo client from serverClient to avoid middleware compilation issues
