+++ src/components/AccountDashboard/AccountDashboard.tsx
@@ -20,7 +20,7 @@ import useMobile from 'hooks/useMobile'
 import { useMyProjectsQuery } from 'hooks/useProjects'
 import { useWalletSignIn } from 'hooks/useWalletSignIn'
 import { useWallet } from 'hooks/Wallet'
-import client from 'lib/apollo/client'
+import { client } from 'lib/apollo/client'
 import { Profile } from 'models/database'
 import Link from 'next/link'
 import { useRouter } from 'next/router'
