+++ src/lib/apollo/client.ts
@@ -8,33 +8,8 @@ import { FunctionsMap, withScalars } from 'apollo-link-scalars'
 import { BigNumber } from 'ethers'
 import { IntrospectionQuery, buildClientSchema } from 'graphql'
 import { isBigNumberish } from 'utils/bigNumbers'
-import { isBrowser } from 'utils/isBrowser'
 import introspectionResult from '../../../graphql.schema.json'
-
-const subgraphUri = () => {
-  let uri: string | undefined
-  if (isBrowser()) {
-    uri = process.env.NEXT_PUBLIC_SUBGRAPH_URL
-    if (!uri) {
-      throw new Error(
-        'NEXT_PUBLIC_SUBGRAPH_URL environment variable not defined',
-      )
-    }
-  } else {
-    uri = process.env.GRAPHQL_SCHEMA_SUBGRAPH_URL
-    if (!uri) {
-      throw new Error(
-        'GRAPHQL_SCHEMA_SUBGRAPH_URL environment variable not defined',
-      )
-    }
-  }
-
-  const url = new URL(uri)
-  if (url.pathname.match(/graphql$/g)) {
-    return url.href.slice(0, url.href.lastIndexOf('/'))
-  }
-  return url.href
-}
+import { subgraphUri } from './subgraphUri'
 
 const typesMap: FunctionsMap = {
   BigInt: {
@@ -63,4 +38,4 @@ const client = new ApolloClient({
   link: ApolloLink.from([scalarsLink, httpLink]),
 })
 
-export default client
+export { client }
