+++ src/pages/api/juicebox/project/[projectHandle].page.ts
@@ -1,5 +1,5 @@
 import { PV_V2 } from 'constants/pv'
-import { paginateDepleteProjectsQueryCall } from 'lib/apollo'
+import { paginateDepleteProjectsQueryCall } from 'lib/apollo/paginateDepleteProjectsQuery'
 import { getLogger } from 'lib/logger'
 import { NextApiRequest, NextApiResponse } from 'next'
